dominoes: Add more tests
First, a small test to point out the necessity of reversing a domino
seems prudent - it will provide a good stepping stone to the more
complex solutions.

Next, we need some tests of disconnected graphs. The rationale is that
there is currently only one test of a graph with no Eulerian cycle, and
that is one where some vertices have odd degree. Many solutions
pre-emptively check that condition, which is good. However, for more
complete coverage it would be wise to include some graphs where all
vertices do have even degree, yet there is still no cycle.

A few current submissions fail this second class of test by declaring
that there is a cycle when in fact there is none.
This includes example.rs in this repository unfortunately. A fix for
that will come in the next commit.

Finally, a few tests to check that solutions aren't relying on the
ordering of the input dominoes.

Again, there is at least one current submission that gives an incorrect
result, in this case declaring there is no cycle when in fact there is
one.