rna-transcription: Remove as_ref test
This was introduced in 7877d6389ed47d8c3e6631274c1c66a0b4c4ec54 back
when the exercises was ported to xrust, but there's no clear reason why
we would want the RNA to be convertible to a string.

* It's not relevant to the problem statement.
* There's no clear use case for why this is something that should be
  required.
* Other languages don't do this.
* Other exercises in xrust don't do this.

Given these reasons, it makes more sense to remove it.

Closes #76