Explain some of Rust's projects, appeal
Fixes #88

The Awesome Rust is a well-maintained and wide-ranging list of Rust
projects. Instead of calling out specific projects in the About.md, it
made sense to link to this list.

I didn't mention the "complier has your back" sentiment from #58, but I
did try to get to the same place by saying that all Rust code will be
memory safe & fast.