dominoes: test singletons that can't be chained
Some solutions may automatically declare a singleton input to be valid
without checking it can actually be chained. Singletons still need to be
checked because a singleton (1, 1) can be chained with itself, but
(1, 2) can't. This adds a test for (1, 2). The test for (1, 1) already
existed.

The example solution needs to be corrected to address this as well.