travis: add an allowed-fail build that fails on warnings
I don't intend to be terribly strict on warnings (so as to actually fail
a build with them) but I think I'd like a way to see them, so I'll let
Travis do the job.

This is motivated by the fact that I introduced warnings in #81 and they
were not caught.