Queen Attack: Replace tuple with ChessPosition
Fixes #118

`Queen::new` now expects a `ChessPosition` instead of a tuple. `ChessPosition::new` returns a Result, indicating if the position is valid or not.

I've added a few tests at the start to capture the behavior of `ChessPosition`.  The tests of Queen have their API usage updated, and have been renamed for readability.