Reorder Problems to (hopefully) improve learning
Fixes #127

This commit follows the problem order I suggested in #127, with one
exception. I've put sublist ahead of allergies. Both use enum, but the
enums in allergies can be more of a hassle, as evidenced by all of the
ways I've seen students implement them.

Sublist does have the disadvantage of also requiring a solution that is
generic over multiple types. But I think the best thing to do is keep an
eye on students that reach this part of the track.

Or the later introduction of new exercises could improve this flow.