Improve vector comparison, error messages
Testing iterator equality seems like a nicer way of comparing these two
vectors. And I'm now returning a useful error message when the vectors
are not equal.

If you do

```rust
assert_eq!(expected.iter().eq(actual.iter()))
```

The error message is just that the two iterators are not equal. The
differing elements are not called out in any way. Not as helpful.