Implement Robot Simulator
Tests come from
https://github.com/exercism/x-common/blob/183934754b1847612809db410a8aeb640678f188/robot-simulator.json

As discussed in https://github.com/exercism/xrust/issues/117 we are
providing a stub implementation so that ignored tests do not fail when a
student hasn't yet implemented the functionality they exercise.

A lot of excellent feedback and help provided by

- jonasbb
- petertseng
- steveklabnik
- Ryman
- Dr-Emann

Full details of the design discussion at

https://github.com/exercism/xrust/pull/146

--

Some details that people may care about:

----

Robots are immutable because:

- Immutability is the default in Rust
- No other problem (that I know of) really features Rust's default immutability
- Immutability is not my natural inclination

I figure if I don't expect immutability then other programmers with an
OO-focused background probably don't expect immutability. So a problem
that forces immutability may make me (and them) think about
immutability. Which I think is good.

Immutability brings other benefits, though they aren't exposed by the
tests. It would be very easy to trace your Robot's path, for example.
Re-winding to a specific point on the path would be trivial. All very
nice, but not part of the test suite.

----

The example code has the `build` function because:

There's still some awkwardness around the `new` function, since it takes
x/y and all the internals use Position.

I have to preserve the `new(x, y, direction)` function signature because
of the tests (and because the tests should follow the example of Queen
Attack, which also passes `x` and `y` in as separate parameters.

But that function becomes a pain once I'm working inside the Robot and I
have a Position.

The new `build` function creates a robot using a Position and Direction,
so it can be easily used by all of the Robot's internals. And `new` now
just wraps around `build`.

This is kind of the exact opposite way I'd normally do this. I'd expect
a `build` function to be the public API, do the data manipulation and
then call a private `new` function.

But if I were to do that here then all of the tests would contain:

```
Robot::build(x, y, &direction)
```

Which I think would be non-idiomatic and weird to students.

----