nucleotide-codons: Use collect
The previous code creates a HashMap, inserts into it, then shrinks it.
I'm not convinced we need to explicitly worry about all that, and the
code is made shorter by simply collecting on the iterator. Don't need to
create a `mut CodonInfo`, can just create one and return it immediately.