parallel-letter-frequency: Use &str with lifetime, not String
The previous version had a comment that explained a need for copying the
`&str` into `Strings`, but this can be avoided.

The disadvantage is now that this function can *only* be invoked with
`&'static str`, which means that it can't (for example) accept arbitrary
user input as part of a command-line program. So it works for our style
of testing (all our tests do indeed use `&'static str`) but is
unrealistic code.