etl: use flat_map
Instead of creating a `BTreeMap` and adding to it, we can just use
`flat_map` on the input and `collect` the result.

Note that we need to use a `move` closure here, otherwise `n` does not
live long enough. The closure would normally only take references to
variables on the stack rather than taking ownership. Using the `move`
closure takes ownership. See
https://doc.rust-lang.org/book/closures.html#move-closures for an
explanation for what is going on.