Implement ocr-numbers
Placement
____

We're putting this exercise before Minesweeper, as suggested here
https://github.com/exercism/xrust/pull/172#issuecomment-236404273

The exercise requires some string splitting/iterating that can be kind
of tricky. Minesweeper's solution requires some similar approaches, so
putting them together makes sense.

Tests
____

This includes all of the tests in the standard test suite, but in a
different order.

As discussed
https://github.com/exercism/xrust/pull/172#issuecomment-236369131

This test order should be smoother. Going step-by-step, The student
will have to implement a function that:

1. Checks line count, returns error
2. Checks line length, returns error
3. Returns "?" if not error
4. Parses each of the decimal numbers
5. Parses a line of more than one number
6. Parses many lines of more than one number

Etc
____

I've also turned off rustfmt for each test.  This lets us keep the
formatting of the input numbers. I don't see a way to turn rustfmt
off at the file or directory level. I've confirmed that currently a
[directory-level ignore](https://github.com/rust-lang-nursery/rustfmt/issues/705)
is not supported. But maybe there's a file-wide directive that I'm
missing?

Thanks to @petertseng and @jonasbb for the help!