Implement Space Age
This implementation follows the common test suite.

We designed the tests to encourage (force?) students to implement
solutions using Traits, and for them to use the From trait to convert
numbers to Durations.

In a "Real World" implementation of a Planetary Calendar you might not
implement this code in this same way. That's fine! Exercism isn't here
to model Real Production Code. It's here to help students learn about
language, design and tradeoffs.

By focusing on Traits this problem will hopefully illustrate how Rust
handles design -- no Inheritance, just Composition via Traits -- and
minimizing code duplication.

We've placed this problem after Queen Attack and Roman Numerals. If the
student has gone through the problems in order then they've covered
Traits (Roman Numerals, and maybe in Queen Attack).  Custom traits may
be new, depending on how they handled Queen Attack.

I think providing a default implementation for a trait function will be
new, but the stub file makes it pretty clear what's going on there.

I don't think it's a big leap to add a trait function that will be
implemented differently for each implementation of Planet. But we'll see
what students make of it.

We're also trying something new, adding a topics.md file that will list
what parts of Rust the problem & solution focus on

Lengthy discussion of how this problem was implemented can be found at
https://github.com/exercism/xrust/pull/177