Implement Bracket Push
Full discussion of this change is at

https://github.com/exercism/xrust/pull/178

Tests
----

Tests follow the canonical json file.

Example
____

The example is a bit of an experiment in Dependency Injection. I'm willing to
admit that it might make this solution worse. But it doesn't affect the
students at all, just the example.

The idea here is that brackets and how they pair is a configuration
detail, not a core part of the "are_balanced" algorithm. So I've changed
Brackets so that it accepts an Option<Vec> of arbitrary brackets. Maybe you
want to match "/" and "\", or "<" and ">". I dunno, I'm not you. Look at
all the neat brackets you can use!

https://en.wikipedia.org/wiki/Bracket

But brackets has a default to fall back on should the dependency not be
provided.

Just using a bare Vec all over the place isn't great either (as far as
Primitive Obsession goes, at least). So I've pulled that part out into
its own named concept -- MatchingBrackets.

MatchingBrackets should know if things are matched, right? So if can
accept 2 unknown brackets and return the ones that are unmatched (which
will always be 0 or 2) , then Brackets doesn't have to concern itself
with what to push back into the array. It'll just shove in whatever
`unmatched` returns.

Currently the `closed_by` function lives in the wrong place. MatchingBrackets is
about a collection of paired brackets, this function is concerned with a
single set of paired brackets. The API for this should be

`bracket.closed_by(other_bracket)`

Instead of the nonsensical `brackets.closed_by(left_bracket,
right_bracket)`.

This will require the extraction of a new struct, though. And this
example was already kind of overdone.

Placing problems
----

I think @petertseng's instincts are right

https://github.com/exercism/xrust/pull/178#issuecomment-242252735

But I moved it before Queen Attack because I think having Queen Attack
followed by Sublist is a good combo.