Handle results, reorder tests, fix errors
Errors
----

I think I caught all of the logical errors in my use of ranges. A second
set of eyes checking that will be much appreciated

Ordering
----

This mostly follows the new canonical_data.json file. I've moved some of
the exception tests up front so that students can quickly handle the
broad edge cases of the problem

- Rolls return results
- Score return results
- A single roll can't be more than 10
- A game needs 10 frames to be OK

The weirder edge cases are saved until the end as I don't think they
need to be handled up front.

Handle results
----

Based on the conversation here
https://github.com/exercism/xrust/pull/213#issuecomment-253384376

I think a bare `is_ok` is fine for all of the rolls that we expect be
ok. We could wrap it in `assert!`, but I think that is noisy as it
indicates we're testing something that we're not actually testing.