Refactoring of example
Other than the very hairy bonus validation code, I'm OK with this
example.

The approach of calling `add_roll` on every frame can be seen as
overkill, but I think it's the frame's responsibility to know what to do
with a roll, not the game's responsibility to decide which frame to add
a roll to. Yeah, it's always going to be the current frame & previous
frame, I realize. But adding it to each frame reads cleaner.