Implement grains
Follows the standard test suite, which is currently waiting for merge.

https://github.com/exercism/x-common/pull/420

I did not include the test of square -1 since that would require the
function signature to be `i32`.

Placement is at the end of the "Introduction" section. It introduces a
couple of concepts that may be new,

- declaring number types
- panic & should_panic

I thought it would be nice to get some exposure to `panic` before
introducing `result` in the Hamming exercise.