Adding a stub src/lib.rs file
This is pretty early in the track, so signatures can be handy. And not
getting a bunch of compilation errors is also nice. I don't think that
this stub forces any particular implementation, though the setting of
type of `s` to `u32` does help the students over a bit of type-casting
hassle.