bowling: propose tests for err semantics
Specifically: if a call to `roll()` is made that results in an error,
**the game state should not be altered**, so that future calls to
`roll()` will proceed as if the erroneous roll **never happened**.

This is only a proposal because there was doubt in #213 about whether
they should be added.

Since it was suggested, after seeing it in my Exercism submission, that
I open an issue for it, I will, because I can let code speak to
accompany my words.

Right now there's no good way to express this kind of semantics in
x-common, but maybe a general discussion on it would be OK to brainstorm
some ways to do it.

TODO:

* The current example solution needs to be fixed.
* It is possible we can add more tests of this type - basically any time
  we expect an `is_err()` from a roll.

I would only want to put in the work to proceed with these if we decide
we *do* want these kinds of tests.