Implement sum-of-multiples
Closes #234

Tests follow the current standard. I have some problems with the current
standard in that almost none of the tests introduce new behaviors or
have useful descriptions. But I'll address that with a PR against
the current standard test suite, not here.

My example implementation is likely poor, but it works.

I'm placing this early in the track because it doesn't really introduce
much Rust-specific behavior beyond the `Vec` being borrowed. The solution
is an algorithmic problem, not a language fluency one.