Update language docs (#233)
Fixes #232

- Update the installation directions to point people to the main page.
This has a few advantages to the other link:

1. It will take them to a language-localized page, if available.
2. It will give them a OS-specific install link.
3. It's kept more up-to-date than The Book.

- Update tests doc

Some of the changes here are just little edits. And I'm mentioning that
some exercises now come with a stub file is `src/lib.rs`.

The compilation errors generated by ignored tests is a frequent source
of confusion, I think. So I documented the easiest way around it,
commenting out the tests.