problems.md: Remove topics
History:

Thanks to the hard work of #167 and #127, we have topics in problems.md
that tell why we ordered problems a certain way.

After #229 (motivated by #179), we copied the topics into config.json,
since now all tracks can take advantage of the "topics" concept.

To reduce duplication, we can remove topics from problems.md now.

What's left in problems.md is a bit of prose explaining why we order our
problems the way they do, and the division of the problems into
sections.

Closes #236