Place Luhn before Largest Series Product
I don't think my solution is that far from what people will do. The
[existing crate](https://lunemec.github.io/rust-luhn/src/luhn2/src/lib.rs.html)
uses another crate to handle digits > 10 after doubling. If people go
down that path their solutions will vary a bit from mine, but not by a
bunch.

So the solution mostly requires a good handle on

- Converting a string to digits
- Working with iteration

The conversion of a string to digits is also part of Largest Series
Product, so we could put the exercise to right before LSP to reinforce
that concept.