Redo the scrabble-score example for the double-letter-value extension.
Multiplying a letter value with a modifying token introduces dependency
between characters of the word. This inter-character dependency makes it
hard to keep the old map+fold method, which, in all its glorious
functional-ness, doesn't like state between its iterations.

We switch to a for-loop, which mutates state kept in the 'score' and
'previous_char_value' variables.
Idiomatic rust prefers for-loops over map() if the primary effect of
the iteration is the side-effect, as is now the case.

Unfortunately, the inter-loop dependency sacrifices (potential) iteration
parallelism. An alternative approach that could conserve parallelism would
be to replace any "X:double" with "XX" and "X:triple" with "XXX".
However, that capability is beyond std's String::replace(), and
introducing the (admittedly awesome) regex-crate just for this is
probably overkill..

The previous HashMap dictionary() is no longer suitable, as the value of
some tokens ('2' and '3') now dynamically depends on the previous char.
Also, Rust's match() is precisely 102% cooler than a HashMap, so use that!