hello-world: simplify to unconditional "Hello, World!"
The introductory exercise should require as little Rust knowledge as
necessary. We'd like to introduce `cargo test` here since it will be
used for all exercises, but introducing Option is too much.

We add a stub in case it is not clear how to return a string. That way,
the student only has to change the string provided.

https://github.com/exercism/x-common/issues/520