grade-school: Expect `Option<Vec<String>>` from grade
Previously, grade was expected to return an `Option<&Vec<String>>`.
Because there's a reference inside the lifetime, any implementing School
would have been required to maintain ownership of some vectors so that
they can be lended out. This unnecessarily restricts the implementation.
For example, some implementations may instead have preferred to use
BTreeSet.

If we expect grade to return `Option<Vec<String>>`, the fact that
ownership is given means that a new vector can be made from whatever
internal representation is used.

Closes #254.