copy nucleotide-codons to protein-translation
As noted in https://github.com/exercism/x-common/issues/268 these two
exercises are the same. In this transitionary period, we deprecate
nucleotide-codons (but keep its files available).

If we wish to replace nucleotide-codons with protein-translation
completely, then a database surgery will be needed to transfer all
submissions from nucleotide-codons to protein-translation.