Fix queen-attack's #[should_panic] parameter
The .unwrap() method does not generate a panic string, so the test
always failed.  The test can be fixed either by keeping the .unwrap()
call and removing the `expected` argument from the attribute, or by
keeping the attribute as-is, and using the .expect() method.

I've picked the former option to remain consistent with other tests (in
this problem and others) that use .unwrap().