difference-of-squares: Test one function at a time
Instead of testing all three functions for 5 and then all three
functions for 100, let's test one function at a time, so that the
student can make sure one function is solid before moving onto the next.

This also better matches the order the tests are listed in in x-common
(though we are free to change the order if we really want).

Honestly, I don't feel strongly about this, so this is just arbitrary.