gigasecond: chrono 0.3
Ever since `rust 1.19` was promoted to beta, Travis builds have been
failing.

1.18:
https://travis-ci.org/exercism/xrust/jobs/237495440

1.19:
https://travis-ci.org/exercism/xrust/jobs/240210670

The errors are multiple errors of the form:

    error: to use a constant of type `format::ParseError` in a pattern, `format::ParseError` must be annotated with `#[derive(PartialEq, Eq)]`
       --> /home/travis/.cargo/registry/src/github.com-1ecc6299db9ec823/chrono-0.2.19/src/format/parsed.rs:476:22
        |
    476 |                 (Err(OUT_OF_RANGE), _) | (_, Err(OUT_OF_RANGE)) => return Err(OUT_OF_RANGE),
        |                      ^^^^^^^^^^^^

That means the necessary commit required to fix it is
https://github.com/chronotope/chrono/commit/0b3218289fe9422ac48631423689fe86658905f9
which appears, at the earliest, in 0.2.23

So, we could choose to stay at 0.2.25 (the latest in 0.2.x), but let's
choose to upgrade to 0.3, since it doesn't seem necessary to hold
ourselves back.