config: rectangles topics: replace lifetimes with algorithm
This has been a long time coming.
Ever since https://github.com/exercism/rust/pull/167#r70376973 lifetimes
should not have been a topic for rectangles.
The reasons were discussed there already:

* The "Putting it All Together" problems do not emphasise Rust language
  features in general.
* The specific feature "lifetimes" is not implied by this problem. A
  sampling of ten submitted solutions shows two solutions with
  lifetimes.

I used the https://github.com/exercism/x-common/issues/834 command to
inspect other tracks' topics to see if we can glean any insight here.

C#/F#: parsing, transforming
Lua: strings, performance, control-flow (if/else, loops), algorithms

Based on these, my recommendation:

A possible topic is string parsing, because we are reading in a string
and trying to make sense of it. However this is not parsing in the usual
sense that one would use a parser for, which usually deals with text.

If we were to declare "string parsing" a topic of rectangles, we would
probably use it for ocr-numbers too (but we do not; the topics for
ocr-numbers are lines, slices, chunks).

"algorithm" is chosen because significant work will likely go into
devising an algorithm for determining the rectangles, since that is not
immediately obvious from the problem description.  Solutions will likely
display a diverse set of approaches.

One might surmise that the alphametics exercise could have this topic as
well, but this commit deals with rectangles only.