Implement exercise crypto-square (#341)
Stub:

It's irritating when you have to puzzle out the intended API from the
test cases; I far prefer it when there's a stub with the required API
pre-prepared, just unimplemented. I've therefore put it in here.

Example:

Pretty proud of this: it's got very few heap allocations, and the
custom iterator indexing should be very fast.