Add a new exercise to teach implementing traits.
Per the comment [here](https://github.com/exercism/rust/pull/347#issuecomment-326798202),
there's continued interest in a Rust-specific exercise focusing on teaching trait
implementation. I've got some ideas for how to structure such an exercise, and want to
stake a claim to flesh them out. That said, I'm very interested in input as to the specifics
of how to go about this.