Add a utility to assist in manually testing an exercise
Instead of having to manually enable tests, copy the examples, and
clean up, this script automates all that.

- Take optional exercise path argument
- Express all paths in terms of exercise path
- Pass further arguments through to cargo
- Use arrays to simplify file/dir preservation and reset
- Use trap to ensure reset occurs on ctrl-c
- Express check-exercises.sh in terms of bin/test-exercise
- Add debug tests to check-exercises in case of remote execution failure
- Move $![deny(warnings)] insertion into test-exercise.
   - Give the feature to people testing the exercise on its own
   - Clean up properly after
- Return the bitwise OR of individual exercises' return values

That last one wants some commenting. If a bunch of values fail in
different ways, this is of limited utility; your return code won't
be 0, and Travis will fail, but you won't get the info you expected
from the return codes. However, this does have some useful properties:
   - if a bunch of things fail in the same way, or only one thing
     fails, you get Cargo's return code back directly
   - you discard less overall information than the naïve implementation
     ( if [ $? != 0 ]; then return_code=1; fi )
   - always returns zero or non-zero appropriately without branching