DO NOT MERGE: Examples for #386
This commit is designed to cause Travis to fail, if #386 is merged
and implemented correctly.

It should check the following exercises, modified by this PR:
  all-your-base
  crypto-square
  decimal
  hello-world
  rectangles
  two-bucket

The following two exercises should be reported as missing README.md:
  all-your-base
  decimal

The followign two exercises should be reported as having out of date READMEs.md:
  crypto-square
  hello-world

The following two exercises should be checked, but pass without further output:
  rectangles
  two-bucket

Note that if Travis runs on these without including the
`_test/ensure-readmes-are-updated.sh` test, it will likely
pass without comment.