forth 1.4.0: Add dedicated tests for case-insensitivity
This commit add tests for case insensitivity to `forth` to better match
the problem description (which says "words are case-insensitive").
Previously there was some testing for case-insensitivity, but it was
limited to `dup`.

Unlike the canonical data (where these are all at the end), these are
interspersed with the words they test.

https://github.com/exercism/problem-specifications/pull/979