Improve test-exercise: sed now more precisely excises #[ignore] lines
Previous implementation simply deleted all lines containing the string
`[ignore]`. This was not desired behavior, because it also deleted
lines which used the attribute inline.

The updated version now removes the string `#[ignore]` and any
trailing whitespace. If the line is blank in consequence, it deletes
the entire line; otherwise it leaves it.