remove `use std::ascii::AsciiExt` for rust 1.23
In rust 1.23, the `AsciiExt` methods are implemented directly on the
concerned types, so attempting to import `AsciiExt` results in an
unused import.

Rust language PR:
https://github.com/rust-lang/rust/pull/44042

PSA:
https://users.rust-lang.org/t/psa-dealing-with-warning-unused-import-std-ascii-asciiext-in-today-s-nightly/13726

This started warning as an unused import on nightly sometime between
2017-11-02 and 2017-11-09:

https://travis-ci.org/exercism/rust/builds/296391447
https://travis-ci.org/exercism/rust/builds/299770403

This started warning on beta exactly when 1.23.0 became Beta (it was
accepted on 1.22.0):

https://travis-ci.org/exercism/rust/builds/303139434
https://travis-ci.org/exercism/rust/builds/306431846

As 1.23 is now stable and our repo disallows warnings on stable, we will
need this.