simple-linked-list: fix hints.md
Closes #413.

I still don't know what the underlying root cause of this error is;
a variety of Markdown renderers have all struggled on the same line,
and each has produced a different output. The one thing all the output
has in common is that it's wrong.

Luckily, we don't need to know what's precisely wrong with this in
order to be able to fix it. This version should be much easier to
render consistently and correctly.