Remove all instances of Cargo.lock and add to .gitignore
One fairly common nit that we encounter is maintaining changes
to Cargo.lock. This file is automatically generated and not required
for the proper functioning of rustc, cargo, or exercism.

Exercism operates on what is fundamentally a library model: students
write libraries, which are tested automatically by the provided
test suites. The official rust guidance is that Cargo.lock should
not be included in the repository for libraries.

As such, this change adds all exercise instances of Cargo.lock to
a project-wide .gitignore, and excises all instances currently
included in the repository.