collatz-conjecture: Use Option<T> instead of Result<T, str>
Since there is only one kind of error that can occur (a non-positive
number), no extra information is given by the `str`, and I do not like
to encourage using `str` as an error type.

I examined a user poll of when to use Result vs Option:
https://www.reddit.com/r/rust/comments/2j0k21/rust_option_vs_result/

One user says to use Option when there is only one failure mode, which
supports the approach taken in this commit.

However, in contrast, there are others who say Result is for any case
when you need to explain the failure. It is possible that we may want to
explain the failure even if there is only one way for a function to
fail.

So the question is whether this error warrants an explanation. Perhaps
any necessary explanation can be produced in the documentation... and
the documentation here does say take any **positive** integer.