perfect-numbers: Use Option<T> instead of Result<T, str>
Since there is only one kind of error that can occur (a non-positive
number), no extra information is given by the `str`, and I do not like
to encourage using `str` as an error type.

I examined a user poll of when to use Result vs Option:
https://www.reddit.com/r/rust/comments/2j0k21/rust_option_vs_result/

One user says to use Option when there is only one failure mode, which
supports the approach taken in this commit.

Same operation as that performed for collatz-conjecture in
https://github.com/exercism/rust/pull/441