check-exercises: Move `echo` before `test-exercise`
Notice that below `test-exercise` there is a check of `$?`. This means
nothing should come after `text-exercise`.

If we fail to do this, warnings will never cause the build to fail (as
we would like them to), because the `$?` is examining the exit code of
`echo`.

We need to move the `echo` (progress dot) that we added in
https://github.com/exercism/rust/pull/446
https://github.com/exercism/rust/pull/446/commits/a3a7b16c0715d03be93e942ef55ec136917ff3b5