dominoes: Use &[T] instead of &Vec<T> for input
The tests create a vector and immediately take a reference to it, doing
nothing else to the vector.
If this were on a single line, Clippy would suggest:
useless use of `vec!`

https://rust-lang-nursery.github.io/rust-clippy/master/index.html#useless_vec

Since these vectors are constructed strictly to express the input
dominoes and are never expected to grow/shrink, it is reasonable to
simply use a slice.