custom-set: Use `&[...]` instead of `vec![...]`
The current tests of CustomSet pass in an owned Vec to `CustomSet::new`.

This actually seems to make sense, as it allows a FromIterator approach
and avoids cloning.

Note that this doesn't obviate the need for the T to be Clone, because
unions will surely require cloning.

The alternative approach expressed by this PR is to state that `new`
will not consume the input, instead cloning its elements.

The example had to be reworked to make this work, so this affects the
interface students are expected to adhere to.

I (the author of this commit) am no longer sure this is a good idea, so
let's have some discussion on it.