react: return erroneous CellID from create_compute
Instead of encouraging `()` as an Err type, we'd prefer to encourage
informative errors. In this case, there is only one failure mode (so we
do not need a separate error type) but there is a parameter (exactly
which cell is invalid) so we'll keep Result to designate the invalid
cell.

As discussed in https://github.com/exercism/rust/issues/444
A subtask of https://github.com/exercism/rust/issues/462