ensure-stubs-compile: Deny warnings in tests
We'd eventually like to deny all warnings from the stubs themselves as
well, but this is a good intermediate state.

Immediate motivation is avoiding an overflowing literal, but I imagine
there are other unforeseen benefits as well.

Note that this required backing up and restoring the tests directory,
since we are editing it.

Closes https://github.com/exercism/rust/issues/461