regenerate all READMEs to remove link to GitHub team
https://github.com/exercism/rust/pull/472 removed the link from the
template because only org members can access the link. This commit
regenerates the README to do the same.

Discussion in https://github.com/exercism/rust/issues/414

It's my belief that this is the only change to the READMEs (no other
README changes from problem-specifications are pending) because upon
reverting #472 and attempting to regenerate, no other changes appear.