react: remove #[allow(unused_variables)]
Note that in this case, we couldn't follow the standard procedure
of adding the variables to the unimplemented block, because the
generic types F and T aren't bounded by Debug. Worse, the closures
expected for F possibly _can't_ be bounded by Debug.

We use named underscore variables as a fallback.