ensure-stubs-compile: continue early if allowed not to compile
Strictly speaking this should have been done ever since #483, but it
wasn't necessary then because touching a file doesn't impose any costs.
But if we were to impose additional work on ensure-stubs-compile, then
we would actually want to do this. This commit paves the way for future
additions to this script.