Add a template to the `accumulate` exercise
With an empty `src/lib.rs` the test suite won't compile until a
learner has written a more or less complete solution. This means they
can't pass the first test until they've understood the whole suite.

This commit adds stub implementations of the `map_function` and
`map_closure` functions so that the test suite compiles as each test
is passed.